+++ tests/test_threedi_api.py
@@ -6,7 +6,12 @@ from mock import AsyncMock
 from threedi_api_client import ThreediApi
 from threedi_api_client.aio.openapi.api_client import ApiClient as AsyncApiClient
 from threedi_api_client.openapi import ApiClient
-from threedi_api_client.openapi.api import V3AlphaApi, V3Api
+from threedi_api_client.openapi.api import V3Api
+
+from threedi_api_client.versions import API_VERSIONS
+
+
+V3AlphaApi = API_VERSIONS['v3-alpha']
 
 
 @pytest.fixture
@@ -17,6 +22,14 @@ def config():
         "THREEDI_API_PASSWORD": "password",
     }
 
+@pytest.fixture
+def token_config():
+    return {
+        "THREEDI_API_HOST": "localhost:8000",
+        "THREEDI_API_USERNAME": "username",
+        "THREEDI_API_ACCESS_TOKEN": "token",
+        "THREEDI_API_REFRESH_TOKEN": "refresh_token"
+    }
 
 @pytest.fixture
 def v3_api(config):
@@ -36,8 +49,28 @@ def test_init_from_env_vars(monkeypatch):
     monkeypatch.setenv("THREEDI_API_HOST", "localhost:8000")
     monkeypatch.setenv("THREEDI_API_USERNAME", "username")
     monkeypatch.setenv("THREEDI_API_PASSWORD", "password")
-    ThreediApi()
+    config = ThreediApi()._api.api_client.configuration
 
+    assert config.username
+    assert config.password
+    assert config.api_key['Authorization'] is None
+    assert config.api_key['refresh'] is None
+
+def test_init_with_tokens(token_config):
+    config = ThreediApi(config=token_config)._api.api_client.configuration
+
+    assert config.username
+    assert config.password is None
+    assert config.api_key['Authorization']
+    assert config.api_key['refresh']
+
+def test_init_with_password_and_token_disallowed(config):
+    config.update(
+        {"THREEDI_API_ACCESS_TOKEN": "token",
+         "THREEDI_API_REFRESH_TOKEN": "refresh_token"}
+    )
+    with pytest.raises(ValueError):
+        ThreediApi(config=config)
 
 def test_init_from_config(config):
     api = ThreediApi(config=config)
@@ -55,6 +88,15 @@ def test_init_missing_config(key, config):
     del config[key]
     with pytest.raises(ValueError):
         ThreediApi(config=config)
+
+
+@pytest.mark.parametrize(
+    "key", ["THREEDI_API_HOST", "THREEDI_API_USERNAME", "THREEDI_API_ACCESS_TOKEN", "THREEDI_API_REFRESH_TOKEN"]
+)
+def test_init_missing_token_config(key, token_config):
+    del token_config[key]
+    with pytest.raises(ValueError):
+        ThreediApi(config=token_config)
 
 
 @pytest.mark.parametrize("suffix", ["/v3.0", "/v2", "/v6/"])